#17 削除機能のhooks化
